+++ changelog.md
@@ -2,11 +2,48 @@
 <b>Full Changelog</b>
 
 ---
+#### v 0.0.3 -- 2024/01/20
+###### New:
+* Make django app (todolist_app)
+* Add Db (models) todolist_app (ToDoList and item) (done makemigrations and migrate)
+* Add todolist\todolist_app\form.py class "form_create"
+
+* API (views)
+    * todolist_app/todolist_create
+    * todolist_app/todolist_read
+    * todolist_app/todolist_delete
+    * todolist_app/todolist_update
+
+* urls
+    * todolist/create
+    * todolist/
+    * todolist/delete/<int:todo_id>/
+
+* Templates 
+    * todolist_app/todolist_create
+    * todolist_app/todolist_read
+    * todolist_app/todolist_edit.html
+
+
+###### Change:
+* API (views)
+    * register/register
+        * Change HttpResponseRedirect('/login')
+    
+* Lainnya :
+    * settings.py 
+        * Add INSTALLED_APPS ['todolist_app.apps.TodolistAppConfig', ]
+    * first_django\urls.py
+        * Add path('todolist/', include("todolist_app.urls")),
+    * register\base.html
+        * Add lokasi sidebar ToDo
+
+---
 #### v 0.0.2 -- 2024/01/20
 ###### New:
 * Make django app (register)
 * Add pip install crispy-bootstrap4
-* Add first_django\register\form.py
+* Add todolist\register\form.py
 
 * API (views)
     * register/base